pbi 44

+++ src/views/Users/ShowUsers.vue
@@ -1,9 +1,10 @@
 <script setup>
 import { ref, onBeforeMount } from "vue";
-import { useRoute } from "vue-router"
-import { deleteUser , getUsers } from "../../Fetch/fetch_user.js"
-import ShowUsers from "../../components/UserComponents/ShowUsers.vue"
-let { params } = useRoute()
+import { useRoute } from "vue-router";
+import { deleteUser, getUsers } from "../../Fetch/fetch_user.js";
+import { checkOwner } from "../../Fetch/fetch_owners";
+import ShowUsers from "../../components/UserComponents/ShowUsers.vue";
+let { params } = useRoute();
 
 const users = ref([]);
 
@@ -13,29 +14,80 @@ onBeforeMount(async () => {
 });
 
 const deleteUsers = async (id) => {
-  const res = await deleteUser(id)
-  res == 200? users.value = users.value.filter((user) => { return user.id != id; }) : ''
-}
+  const categoryOWners = await checkOwner(id);
+  if (categoryOWners.length > 0) {
+    Swal.fire({
+      title: "This lecturer is an owner in the category",
+      text: "If you delete this user, ownership will also be deleted",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Done",
+    }).then(async(res) => {
+      if (res.isConfirmed) {
+        const res = await deleteUser(id);
+        res == 200
+          ? (users.value = users.value.filter((user) => {
+              return user.id != id;
+            }))
+          : "";
+      }
+    });
+  } else {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const res = await deleteUser(id);
+        res == 200
+          ? (users.value = users.value.filter((user) => {
+              return user.id != id;
+            }))
+          : "";
+      } else {
+        console.log("Failed to execute! This's user hasn't been deleted.");
+      }
+    });
+  }
+  // console.log(categoryOWners);
+  // const res = await deleteUser(id);
+  // res == 200
+  //   ? (users.value = users.value.filter((user) => {
+  //       return user.id != id;
+  //     }))
+  //   : "";
+};
 
 const fillers = () => {
-  if (params.roles != 'all') {
-    users.value = users.value.filter(user => user.role == params.roles)
+  if (params.roles != "all") {
+    users.value = users.value.filter((user) => user.role == params.roles);
   }
-}
-const render = !(localStorage.getItem("accessToken") == null)
-
+};
+const render = !(localStorage.getItem("accessToken") == null);
 </script>
 
 <template>
-  <ShowUsers :users="users" @deleteUsers="deleteUsers" v-if="render"></ShowUsers>
+  <ShowUsers
+    :users="users"
+    @deleteUsers="deleteUsers"
+    v-if="render"
+  ></ShowUsers>
   <h1 id="pleaslogin" v-else>Please Login</h1>
 </template>
 
 <style scoped>
-  #pleaslogin{
-    text-align: center;
-    margin-top: 20em;
-  }
+#pleaslogin {
+  text-align: center;
+  margin-top: 20em;
+}
+
 .detail {
   margin-left: 5%;
   font-size: 100%;
